refactor(codegen): reorganize flat layout into sub-packages
Move modules into three sub-packages matching the architecture layers:

- extraction/ (14 modules): type analysis, specs, extractors, constraints
- layout/ (2 modules): module layout, type collection
- markdown/ (6 modules + templates): pipeline, renderer, type formatting,
  links, paths, reverse references

Three modules renamed to drop redundant prefixes:
  field_constraint_description → extraction/field_constraints
  model_constraint_description → extraction/model_constraints
  example_loader → extraction/examples

Templates flattened from templates/markdown/ to markdown/templates/.
