+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/case_conversion.py
