+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/enum_extraction.py
