+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/examples.py
