+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/model_constraints.py
