+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/primitive_extraction.py
