+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/pydantic_extraction.py
