+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/specs.py
