+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/type_analyzer.py
