+++ packages/overture-schema-codegen/src/overture/schema/codegen/extraction/union_extraction.py
