+++ packages/overture-schema-codegen/src/overture/schema/codegen/markdown/templates/enum.md.jinja2
