+++ packages/overture-schema-codegen/src/overture/schema/codegen/markdown/templates/geometry.md.jinja2
