+++ packages/overture-schema-codegen/src/overture/schema/codegen/markdown/templates/newtype.md.jinja2
