+++ packages/overture-schema-codegen/src/overture/schema/codegen/markdown/templates/primitives.md.jinja2
