+++ packages/overture-schema-codegen/src/overture/schema/codegen/markdown/templates/pydantic_type.md.jinja2
